Refactor app architecture.
