+++ src/main/kotlin/br/com/tagliaferrodev/dextra/pottertest/character/CharacterRepository.kt
@@ -1,0 +1,9 @@
+package br.com.tagliaferrodev.dextra.pottertest.character
+
+import br.com.tagliaferrodev.dextra.pottertest.character.domain.Character
+import org.springframework.data.jpa.repository.JpaRepository
+import org.springframework.stereotype.Repository
+import java.util.*
+
+@Repository
+interface CharacterRepository : JpaRepository<Character, UUID>
