+++ src/main/kotlin/br/com/tagliaferrodev/dextra/pottertest/character/CharacterService.kt
@@ -1,0 +1,7 @@
+package br.com.tagliaferrodev.dextra.pottertest.character
+
+import org.springframework.stereotype.Service
+
+@Service
+class CharacterService {
+}
