+++ src/main/kotlin/br/com/tagliaferrodev/dextra/pottertest/character/domain/Role.kt
@@ -1,0 +1,7 @@
+package br.com.tagliaferrodev.dextra.pottertest.character.domain
+
+enum class Role {
+    STUDENT,
+    CLASS_LEADER,
+    TEACHER
+}
