--- src/main/kotlin/br/com/tagliaferrodev/dextra/pottertest/enum/Role.kt
@@ -1,7 +1,0 @@
-package br.com.tagliaferrodev.dextra.pottertest.enum
-
-enum class Role {
-    STUDENT,
-    CLASS_LEADER,
-    TEACHER
-}
